lib::audio; track written samples
In order for AudioTiming to make predictions that don't depend on being
synchronized with the ring state, it needs to know how many samples were
consumed at the instant of the given snapshot.

This enables consistent extrapoltion of a read goal just by knowing the quantum
size and the variance.

+++ mutate-lib/src/audio/mod.rs
@@ -655,7 +655,7 @@ impl AudioProducer {
 
         // Update the timing data
         // XXX this synchronization story seems silly for the device import callback
-        let snapshot = conn.timing.observe(arrived, written);
+        let snapshot = conn.timing.observe(arrived, written as u64);
         let mut audio_timing = conn.lock.lock()?;
         *audio_timing = snapshot;
